Fix the learner’s FP16 policy-mask overflow by promoting gathered legal logits to FP32 before masking and log-softmax. Add regression coverage for half-precision policy logits to prevent the first training step from crashing.

+++ fisher_ai/trainer.py
@@ -107,7 +107,7 @@ class AlphaZeroTrainer:
         value_normalizer,
     ):
         policy_logits, predicted_values = self.model(states)
-        legal_logits = policy_logits.gather(1, legal_actions)
+        legal_logits = policy_logits.gather(1, legal_actions).float()
         legal_logits = legal_logits.masked_fill(~legal_mask, -1e9)
         log_policy = F.log_softmax(legal_logits, dim=1)
